pyproject.toml
[build-system]
requires = ["setuptools>=68", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "xtrader"
version = "0.1.0"
description = "Research-oriented quantitative trading toolkit focusing on Bitget data ingestion."
readme = "README.md"
requires-python = ">=3.11"
authors = [{ name = "XTrader Authors" }]
license = { text = "Proprietary" }
dependencies = [
  "httpx>=0.27",
  "websockets>=12.0",
  "pydantic>=2.6",
  "pydantic-settings>=2.2",
  "python-dotenv>=1.0",
  "pyyaml>=6.0",
  "pandas>=2.2",
  "numpy>=1.26",
  "pyarrow>=15.0",
  "typing-extensions>=4.10",
  "tenacity>=8.2",
  "rich>=13.7"
]

[project.optional-dependencies]
dev = [
  "pytest>=8.1",
  "pytest-asyncio>=0.23",
  "respx>=0.20",
  "coverage>=7.4",
  "pre-commit>=3.6",
  "black>=24.2",
  "ruff>=0.3",
  "isort>=5.13",
  "mypy>=1.8",
  "types-pyyaml>=6.0"
]
ml = [
  "torch>=2.2"
]

[project.urls]
Homepage = "https://example.com/xtrader"
Repository = "https://example.com/xtrader.git"

[tool.setuptools]
package-dir = { "" = "src" }

[tool.setuptools.packages.find]
where = ["src"]

[tool.black]
target-version = ["py311"]
line-length = 100

[tool.isort]
profile = "black"
line_length = 100

[tool.ruff]
line-length = 100
target-version = "py311"
select = ["E", "F", "I", "N", "UP", "B", "S", "W"]
ignore = ["S101"]
exclude = ["docs", "data"]

[tool.ruff.lint.per-file-ignores]
"tests/**/*" = ["S101", "S311"]

[tool.mypy]
python_version = "3.11"
warn_unused_configs = true
warn_return_any = true
warn_unused_ignores = true
strict_optional = true
ignore_missing_imports = true
packages = ["xtrader"]
